+++ lib/contentful.rb
@@ -1,3 +1,4 @@
 require 'contentful/version'
 require 'contentful/support'
 require 'contentful/client'
+require 'contentful/resource/custom_resource'