+++ lib/contentful/entry.rb
@@ -8,5 +8,41 @@ module Contentful
     include Contentful::Resource
     include Contentful::Resource::SystemProperties
     include Contentful::Resource::Fields
+
+    def marshal_dump
+      raw_with_links
+    end
+
+    def marshal_load(raw_object)
+      @properties = extract_from_object(raw_object, :property, self.class.property_coercions.keys)
+      @sys = raw_object.key?('sys') ? extract_from_object(raw_object['sys'], :sys) : {}
+      extract_fields_from_object!(raw_object)
+    end
+
+    def raw_with_links
+      links = properties.keys.select { |property| is_known_link?(property) }
+      processed_raw = Marshal.load(Marshal.dump(raw)) # Deep Copy
+      raw['fields'].each do |k, v|
+        processed_raw['fields'][k] = links.include?(k.to_sym) ? self.send(snakify(k)) : v
+      end
+
+      processed_raw
+    end
+
+    private
+
+    def is_known_link?(name)
+      field_name = name.to_sym
+      is_known_link ||= is_known_contentful_object?(fields[field_name])
+      is_known_link ||= fields[field_name].is_a?(Enumerable) && is_known_contentful_object?(fields[field_name].first)
+    end
+
+    def is_known_contentful_object?(object)
+      (object.is_a?(Contentful::Entry) || object.is_a?(Contentful::Asset))
+    end
+
+    def snakify(name)
+      Contentful::Support.snakify(name).to_sym
+    end
   end
 end